+++ src/classes/worker.ts
@@ -3,10 +3,11 @@ import { Redis } from 'ioredis';
 import * as path from 'path';
 import { v4 } from 'uuid';
 import {
-  Processor,
-  WorkerOptions,
   GetNextJobOptions,
+  JobJsonRaw,
+  Processor,
   RedisClient,
+  WorkerOptions,
 } from '../interfaces';
 import {
   clientCommandMessageReg,
@@ -18,7 +19,7 @@ import {
 import { QueueBase } from './queue-base';
 import { Repeat } from './repeat';
 import { ChildPool } from './child-pool';
-import { Job, JobJsonRaw } from './job';
+import { Job } from './job';
 import { RedisConnection } from './redis-connection';
 import sandbox from './sandbox';
 import { Scripts } from './scripts';
@@ -27,24 +28,34 @@ import { TimerManager } from './timer-manager';
 // note: sandboxed processors would also like to define concurrency per process
 // for better resource utilization.
 
-export interface WorkerDeclaration {
+export interface WorkerListener {
   /**
    * Listen to 'active' event.
    *
    * This event is triggered when a job enters the 'active' state.
-   *
-   * @param event -
    */
-  on(event: 'active', listener: (job: Job, prev: string) => void): this;
+  active: (job: Job, prev: string) => void;
+
+  /**
+   * Listen to 'closing' event.
+   *
+   * This event is triggered when the worker is closed.
+   */
+  closed: () => void;
+
+  /**
+   * Listen to 'closing' event.
+   *
+   * This event is triggered when the worker is closing.
+   */
+  closing: (msg: string) => void;
 
   /**
    * Listen to 'completed' event.
    *
    * This event is triggered when a job has successfully completed.
-   *
-   * @param event -
    */
-  on(event: 'completed', listener: (job: Job) => void): this;
+  completed: (job: Job, result: any, prev: string) => void;
 
   /**
    * Listen to 'drained' event.
@@ -52,28 +63,29 @@ export interface WorkerDeclaration {
    * This event is triggered when the queue has drained the waiting list.
    * Note that there could still be delayed jobs waiting their timers to expire
    * and this event will still be triggered as long as the waiting list has emptied.
-   *
-   * @param event -
    */
-  on(event: 'drained', listener: () => void): this;
+  drained: () => void;
 
   /**
    * Listen to 'error' event.
    *
    * This event is triggered when an error is throw.
-   *
-   * @param event -
    */
-  on(event: 'error', listener: (failedReason: Error) => void): this;
+  error: (failedReason: Error) => void;
 
   /**
    * Listen to 'failed' event.
    *
    * This event is triggered when a job has thrown an exception.
-   *
-   * @param event -
    */
-  on(event: 'failed', listener: (job: Job, error: Error) => void): this;
+  failed: (job: Job, error: Error, prev: string) => void;
+
+  /**
+   * Listen to 'paused' event.
+   *
+   * This event is triggered when the queue is paused.
+   */
+  paused: () => void;
 
   /**
    * Listen to 'progress' event.
@@ -82,14 +94,15 @@ export interface WorkerDeclaration {
    * Job##updateProgress() method is called. This is useful to notify
    * progress or any other data from within a processor to the rest of the
    * world.
-   *
-   * @param event -
    */
-  on(
-    event: 'progress',
-    listener: (job: Job, progress: number | object) => void,
-  ): this;
-  on(event: string, listener: Function): this;
+  progress: (job: Job, progress: number | object) => void;
+
+  /**
+   * Listen to 'resumed' event.
+   *
+   * This event is triggered when the queue is resumed.
+   */
+  resumed: () => void;
 }
 
 /**
@@ -99,15 +112,13 @@ export interface WorkerDeclaration {
  *
  */
 export class Worker<
-    DataType = any,
-    ResultType = any,
-    NameType extends string = string,
-  >
-  extends QueueBase
-  implements WorkerDeclaration
-{
+  DataType = any,
+  ResultType = any,
+  NameType extends string = string,
+> extends QueueBase {
   opts: WorkerOptions;
 
+  public readonly name: string;
   private drained: boolean;
   private waiting = false;
   private running = false;
@@ -128,16 +139,18 @@ export class Worker<
     string
   >;
   constructor(
-    name: string,
+    public readonly queueName: string,
     processor?: string | Processor<DataType, ResultType, NameType>,
     opts: WorkerOptions = {},
     Connection?: typeof RedisConnection,
   ) {
     super(
-      name,
+      queueName,
       { ...opts, sharedConnection: isRedisInstance(opts.connection) },
       Connection,
     );
+
+    this.name = this.opts?.name;
 
     this.opts = {
       drainDelay: 5,
@@ -155,7 +168,7 @@ export class Worker<
         ? (<Redis>opts.connection).duplicate()
         : opts.connection,
     );
-    this.blockingConnection.on('error', this.emit.bind(this, 'error'));
+    this.blockingConnection.on('error', error => this.emit('error', error));
 
     if (processor) {
       if (typeof processor === 'function') {
@@ -175,7 +188,7 @@ export class Worker<
         try {
           fs.statSync(masterFile); // would throw if file not exists
         } catch (_) {
-          masterFile = path.join(process.cwd(), 'dist/classes/master.js');
+          masterFile = path.join(process.cwd(), 'dist/cjs/classes/master.js');
           fs.statSync(masterFile);
         }
 
@@ -191,6 +204,37 @@ export class Worker<
     this.on('error', err => console.error(err));
   }
 
+  emit<U extends keyof WorkerListener>(
+    event: U,
+    ...args: Parameters<WorkerListener[U]>
+  ): boolean {
+    return super.emit(event, ...args);
+  }
+
+  off<U extends keyof WorkerListener>(
+    eventName: U,
+    listener: WorkerListener[U],
+  ): this {
+    super.off(eventName, listener);
+    return this;
+  }
+
+  on<U extends keyof WorkerListener>(
+    event: U,
+    listener: WorkerListener[U],
+  ): this {
+    super.on(event, listener);
+    return this;
+  }
+
+  once<U extends keyof WorkerListener>(
+    event: U,
+    listener: WorkerListener[U],
+  ): this {
+    super.once(event, listener);
+    return this;
+  }
+
   protected callProcessJob(
     job: Job<DataType, ResultType, NameType>,
     token: string,
@@ -198,8 +242,15 @@ export class Worker<
     return this.processFn(job, token);
   }
 
-  protected createJob(data: JobJsonRaw, jobId: string) {
-    return Job.fromJSON(this, data, jobId);
+  protected createJob(
+    data: JobJsonRaw,
+    jobId: string,
+  ): Job<DataType, ResultType, NameType> {
+    return Job.fromJSON(this, data, jobId) as Job<
+      DataType,
+      ResultType,
+      NameType
+    >;
   }
 
   /**
@@ -217,7 +268,7 @@ export class Worker<
     return new Promise<Repeat>(async resolve => {
       if (!this._repeat) {
         const connection = await this.client;
-        this._repeat = new Repeat(this.name, {
+        this._repeat = new Repeat(this.queueName, {
           ...this.opts,
           connection,
         });
@@ -264,8 +315,9 @@ export class Worker<
           while (!this.closing) {
             if (processing.size < opts.concurrency) {
               const token = tokens.pop();
+
               processing.set(
-                this.retryIfFailed<Job<any, any, string>>(
+                this.retryIfFailed<Job<DataType, ResultType, NameType>>(
                   () => this.getNextJob(token),
                   this.opts.runRetryDelay,
                 ),
@@ -290,7 +342,7 @@ export class Worker<
             if (job) {
               // reuse same token if next job is available to process
               processing.set(
-                this.retryIfFailed<void | Job<any, any, string>>(
+                this.retryIfFailed<void | Job<DataType, ResultType, NameType>>(
                   () => this.processJob(job, token),
                   this.opts.runRetryDelay,
                 ),
@@ -353,15 +405,12 @@ export class Worker<
     }
   }
 
-  protected async moveToActive(token: string, jobId?: string) {
-    try {
-      const [jobData, id] = await Scripts.moveToActive(this, token, jobId);
-      return await this.nextJobFromJobData(jobData, id);
-    } catch (error) {
-      if (isNotConnectionError(<Error>error)) {
-        this.emit('error', error);
-      }
-    }
+  protected async moveToActive(
+    token: string,
+    jobId?: string,
+  ): Promise<Job<DataType, ResultType, NameType>> {
+    const [jobData, id] = await Scripts.moveToActive(this, token, jobId);
+    return this.nextJobFromJobData(jobData, id);
   }
 
   private async waitForJob() {
@@ -389,7 +438,7 @@ export class Worker<
       );
     } catch (error) {
       if (isNotConnectionError(<Error>error)) {
-        this.emit('error', error);
+        this.emit('error', <Error>error);
       }
       await this.delay();
     } finally {
@@ -409,7 +458,7 @@ export class Worker<
   protected async nextJobFromJobData(
     jobData?: JobJsonRaw | number,
     jobId?: string,
-  ): Promise<Job<any, any, string>> {
+  ): Promise<Job<DataType, ResultType, NameType>> {
     // NOTE: This is not really optimal in all cases since a new job would could arrive at the wait
     // list and this worker will not start processing it directly.
     // Best would be to emit drain and block for rateKeyExpirationTime
@@ -445,7 +494,7 @@ export class Worker<
   async processJob(
     job: Job<DataType, ResultType, NameType>,
     token: string,
-  ): Promise<void | Job<any, any, string>> {
+  ): Promise<void | Job<DataType, ResultType, NameType>> {
     if (!job || this.closing || this.paused) {
       return;
     }
@@ -503,10 +552,10 @@ export class Worker<
 
     const handleFailed = async (err: Error) => {
       try {
-        const failed = await job.moveToFailed(err, token);
+        await job.moveToFailed(err, token);
         this.emit('failed', job, err, 'active');
       } catch (err) {
-        this.emit('error', err);
+        this.emit('error', <Error>err);
         // It probably means that the job has lost the lock before completion
         // The QueueScheduler will (or already has) moved the job back
         // to the waiting list (as stalled)
@@ -548,7 +597,7 @@ export class Worker<
    *
    * Resumes processing of this worker (if paused).
    */
-  resume() {
+  resume(): void {
     if (this.resumeWorker) {
       this.resumeWorker();
       this.emit('resumed');
@@ -649,6 +698,7 @@ export class Worker<
       try {
         return await fn();
       } catch (err) {
+        this.emit('error', <Error>err);
         if (delayInMs) {
           await delay(delayInMs);
         } else {
@@ -656,5 +706,16 @@ export class Worker<
         }
       }
     } while (retry);
+  }
+
+  protected base64Name(): string {
+    return Buffer.from(this.queueName).toString('base64');
+  }
+
+  protected clientName(): string {
+    const queueNameBase64 = this.base64Name();
+    return `${this.opts.prefix}:${
+      this.name ? `${queueNameBase64}:${this.name}` : queueNameBase64
+    }`;
   }
 }