+++ tests/fixtures/fixture_processor.cjs
@@ -1,4 +1,3 @@
-
 /**
  * A processor file to be used in tests.
  *
@@ -7,7 +6,7 @@
 
 const delay = require('./delay');
 
-module.exports = function(/*job*/) {
+module.exports = function (/*job*/) {
   return delay(500).then(() => {
     return 42;
   });