+++ tests/test_job.ts
@@ -109,6 +109,16 @@ describe('Job', function () {
         );
       });
     });
+
+    describe('when delay and repeat options are provided', () => {
+      it('throws an error', async () => {
+        const data = { foo: 'bar' };
+        const opts = { repeat: { every: 200 }, delay: 1000 };
+        await expect(Job.create(queue, 'test', data, opts)).to.be.rejectedWith(
+          'Delay and repeat options could not be used together',
+        );
+      });
+    });
   });
 
   describe('JSON.stringify', () => {
@@ -253,6 +263,16 @@ describe('Job', function () {
       await job.updateProgress({ total: 120, completed: 40 });
       const storedJob = await Job.fromId(queue, job.id);
       expect(storedJob.progress).to.eql({ total: 120, completed: 40 });
+    });
+
+    describe('when job is removed', () => {
+      it('throws error', async function () {
+        const job = await Job.create(queue, 'test', { foo: 'bar' });
+        await job.remove();
+        await expect(
+          job.updateProgress({ total: 120, completed: 40 }),
+        ).to.be.rejectedWith(`Missing key for job ${job.id}. updateProgress`);
+      });
     });
   });
 
@@ -875,9 +895,16 @@ describe('Job', function () {
 
     describe('when job was added with removeOnComplete', async () => {
       it('rejects with missing key for job message', async function () {
-        const worker = new Worker(queueName, async job => 'qux', {
-          connection,
-        });
+        const worker = new Worker(
+          queueName,
+          async job => {
+            await delay(100);
+            return 'qux';
+          },
+          {
+            connection,
+          },
+        );
         await worker.waitUntilReady();
 
         const completed = new Promise<void>((resolve, reject) => {