+++ tests/test_worker.ts
@@ -36,47 +36,69 @@ describe('workers', function () {
     await removeAllQueueData(new IORedis(), queueName);
   });
 
-  it('should get all workers for this queue', async function () {
-    const worker = new Worker(queueName, async job => {}, { connection });
-    await worker.waitUntilReady();
-    worker.run();
-    await delay(10);
+  describe('when .getWorkers', () => {
+    it('gets all workers for this queue', async function () {
+      const worker = new Worker(queueName, async job => {}, { connection });
+      await worker.waitUntilReady();
+      worker.run();
+      await delay(10);
 
-    const workers = await queue.getWorkers();
-    expect(workers).to.have.length(1);
+      const workers = await queue.getWorkers();
+      expect(workers).to.have.length(1);
 
-    const worker2 = new Worker(queueName, async job => {}, { connection });
-    await worker2.waitUntilReady();
-    worker2.run();
-    await delay(10);
+      const worker2 = new Worker(queueName, async job => {}, { connection });
+      await worker2.waitUntilReady();
+      worker2.run();
+      await delay(10);
 
-    const nextWorkers = await queue.getWorkers();
-    expect(nextWorkers).to.have.length(2);
+      const nextWorkers = await queue.getWorkers();
+      expect(nextWorkers).to.have.length(2);
+      //console.log(nextWorkers,Buffer.from(queueName).toString('base64'));
+      expect(nextWorkers[0].name).to.equal(queueName);
 
-    await worker.close();
-    await worker2.close();
-  });
+      await worker.close();
+      await worker2.close();
+    });
 
-  it('should get only workers related only to one queue', async function () {
-    const queueName2 = `${queueName}2`;
-    const queue2 = new Queue(queueName2, { connection });
-    const worker = new Worker(queueName, async job => {}, { connection });
-    const worker2 = new Worker(queueName2, async job => {}, { connection });
-    await worker.waitUntilReady();
-    await worker2.waitUntilReady();
-    worker.run();
-    worker2.run();
+    it('gets worker with provided name for this queue', async function () {
+      const workerName = `test-${v4()}`;
+      const worker = new Worker(queueName, async job => {}, {
+        connection,
+        name: workerName,
+      });
+      await worker.waitUntilReady();
+      worker.run();
+      await delay(10);
 
-    const workers = await queue.getWorkers();
-    expect(workers).to.have.length(1);
+      const workers = await queue.getWorkers();
+      expect(workers).to.have.length(1);
 
-    const workers2 = await queue2.getWorkers();
-    expect(workers2).to.have.length(1);
+      expect(workers[0].name).to.equal(workerName);
 
-    await queue2.close();
-    await worker.close();
-    await worker2.close();
-    await removeAllQueueData(new IORedis(), queueName2);
+      await worker.close();
+    });
+
+    it('gets only workers related only to one queue', async function () {
+      const queueName2 = `${queueName}2`;
+      const queue2 = new Queue(queueName2, { connection });
+      const worker = new Worker(queueName, async job => {}, { connection });
+      const worker2 = new Worker(queueName2, async job => {}, { connection });
+      await worker.waitUntilReady();
+      await worker2.waitUntilReady();
+      worker.run();
+      worker2.run();
+
+      const workers = await queue.getWorkers();
+      expect(workers).to.have.length(1);
+
+      const workers2 = await queue2.getWorkers();
+      expect(workers2).to.have.length(1);
+
+      await queue2.close();
+      await worker.close();
+      await worker2.close();
+      await removeAllQueueData(new IORedis(), queueName2);
+    });
   });
 
   describe('when closing a worker', () => {
@@ -151,6 +173,55 @@ describe('workers', function () {
       const count = await queue.getJobCounts('active', 'completed');
       expect(count.active).to.be.eq(0);
       expect(count.completed).to.be.eq(1);
+    });
+  });
+
+  describe('when sharing connection', () => {
+    it('should not fail', async () => {
+      const queueName2 = `test-${v4()}`;
+
+      const connection = new IORedis({
+        host: 'localhost',
+        maxRetriesPerRequest: null,
+        enableReadyCheck: false,
+      });
+
+      const queue1 = new Queue(queueName2, { connection });
+
+      let counter = 1;
+      const maxJobs = 35;
+
+      let processor;
+      const processing = new Promise<void>((resolve, reject) => {
+        processor = async (job: Job) => {
+          try {
+            expect(job.data.num).to.be.equal(counter);
+            expect(job.data.foo).to.be.equal('bar');
+            if (counter === maxJobs) {
+              resolve();
+            }
+            counter++;
+          } catch (err) {
+            reject(err);
+          }
+        };
+      });
+
+      const worker = new Worker(queueName2, processor, { connection });
+      await worker.waitUntilReady();
+
+      worker.run();
+
+      for (let i = 1; i <= maxJobs; i++) {
+        await queue1.add('test', { foo: 'bar', num: i });
+      }
+
+      await processing;
+      expect(worker.isRunning()).to.be.equal(true);
+
+      await worker.close();
+      await queue1.close();
+      await removeAllQueueData(new IORedis(), queueName2);
     });
   });
 
@@ -1954,8 +2025,10 @@ describe('workers', function () {
         });
       });
 
-      const completing = new Promise<void>((resolve, reject) => {
-        worker.on('completed', resolve);
+      const completing = new Promise<void>((resolve, _reject) => {
+        worker.on('completed', async () => {
+          resolve();
+        });
       });
 
       queueScheduler.run();
@@ -2011,8 +2084,10 @@ describe('workers', function () {
         });
       });
 
-      const completing = new Promise<void>((resolve, reject) => {
-        worker.on('completed', resolve);
+      const completing = new Promise<void>((resolve, _reject) => {
+        worker.on('completed', async () => {
+          resolve();
+        });
       });
 
       queueScheduler.run();
@@ -2042,7 +2117,7 @@ describe('workers', function () {
     it('should not retry a job that is active', async () => {
       const worker = new Worker(
         queueName,
-        async job => {
+        async () => {
           await delay(500);
         },
         { connection },